# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Ball": {
        "description": "Who acts next. Required while a task is open; null only when closed.",
        "enum": [
          "agent",
          "human",
          "external"
        ],
        "title": "Ball",
        "type": "string"
      },
...
      "ClaimRequest": {
        "description": "An agent takes ownership of a ready task.",
        "properties": {
          "agent": {
            "description": "Actor id claiming the task.",
            "title": "Agent",
            "type": "string"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Caller-generated UUID. Resending the same request with the same id replays the original result instead of writing again; reusing it for a different request is a conflict and writes nothing.",
            "title": "Operation Id"
          }
        },
        "required": [
          "agent"
        ],
        "title": "ClaimRequest",
        "type": "object"
      },
...
      "DispatchRequestBody": {
        "description": "Ask AgentJobs to start an agent on this task.\n\nThere is no ``actor`` field, and that absence is the design. The actor recorded on a\ndispatch is the author of the log entry that *caused* it, not whoever posted the\nrequest -- otherwise the human-clocked rule would be satisfied by a caller naming a\nhuman, which is not evidence of anything.",
        "properties": {
          "caused_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Log entry id authorising this dispatch. Defaults to the newest entry. Its actor must be a configured human.",
            "title": "Caused By"
          }
        },
        "title": "DispatchRequestBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Lifecycle": {
        "description": "Where a task is in its life (design doc section 3).",
        "enum": [
          "draft",
          "ready",
          "active",
          "closed"
        ],
        "title": "Lifecycle",
        "type": "string"
      },
...
      "Priority": {
        "description": "Relative urgency.",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Priority",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "REST API for interacting with AgentJobs tasks, including task management, status tracking, prompt coordination, and search.",
    "title": "AgentJobs API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/tasks": {
      "get": {
        "description": "List tasks filtered along the state axes.\n\n``?ball=human`` is the human inbox: everything waiting on a person, each row\ncarrying its ``ball_prompt``. ``?ball=external`` is the blocked list.\n``?parent=task-063-schema-v2`` is one umbrella's children.",
        "operationId": "list_tasks_api_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "lifecycle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Lifecycle"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lifecycle"
            }
          },
          {
            "in": "query",
            "name": "ball",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Ball"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ball"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Priority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "description": "Return only the children of this umbrella task.",
            "in": "query",
            "name": "parent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return only the children of this umbrella task.",
              "title": "Parent"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskRead-Output"
                  },
                  "title": "Response List Tasks Api Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tasks",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "description": "Create a new task record.\n\nWith an operation_id the create runs under the project-wide creation lock and a\nretry resolves to the task the first attempt made, rather than producing a second\none with a different generated id.",
        "operationId": "create_task_api_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Task",
        "tags": [
          "tasks"
        ]
      }
    },
...
    "/api/tasks/{task_id}/claim": {
      "post": {
        "description": "Claim a ready task: one winner, everyone else gets a 409.",
        "operationId": "claim_task_api_tasks__task_id__claim_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "description": "Return a MutationResult with replayed/warnings instead of the bare task. Defaults to false, so existing callers see no change.",
            "in": "query",
            "name": "envelope",
            "required": false,
            "schema": {
              "default": false,
              "description": "Return a MutationResult with replayed/warnings instead of the bare task. Defaults to false, so existing callers see no change.",
              "title": "Envelope",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/MutationResult-Output"
                    },
                    {
                      "$ref": "#/components/schemas/Task"
                    }
                  ],
                  "title": "Response Claim Task Api Tasks  Task Id  Claim Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Task",
        "tags": [
          "status"
        ]
      }
    },
    "/api/tasks/{task_id}/close": {
      "post": {
        "description": "End the task with an outcome.",
        "operationId": "close_task_api_tasks__task_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "description": "Return a MutationResult with replayed/warnings instead of the bare task. Defaults to false, so existing callers see no change.",
            "in": "query",
            "name": "envelope",
            "required": false,
            "schema": {
              "default": false,
              "description": "Return a MutationResult with replayed/warnings instead of the bare task. Defaults to false, so existing callers see no change.",
              "title": "Envelope",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/MutationResult-Output"
                    },
                    {
                      "$ref": "#/components/schemas/Task"
                    }
                  ],
                  "title": "Response Close Task Api Tasks  Task Id  Close Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close Task",
        "tags": [
          "status"
        ]
      }
    },
...
    "/api/tasks/{task_id}/dispatch": {
      "post": {
        "description": "Start an agent on this task.\n\nDeliberately **not** merged into any approval endpoint. Approving means \"I agree\";\ndispatching means \"spend money now\", and collapsing the two would turn every\napproval into an implicit purchase (design decision D1).\n\n202 rather than 200: the run has started, and how it ends arrives later as\n``dispatch_result`` entries on the task, not in this response.",
        "operationId": "dispatch_task_endpoint_api_tasks__task_id__dispatch_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchRequestBody",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dispatch Task Endpoint",
        "tags": [
          "status"
        ]
      }
    },
    "/api/tasks/{task_id}/handoff": {
      "post": {
        "description": "Move the ball, with its ask.",
        "operationId": "handoff_task_api_tasks__task_id__handoff_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "description": "Return a MutationResult with replayed/warnings instead of the bare task. Defaults to false, so existing callers see no change.",
            "in": "query",
            "name": "envelope",
            "required": false,
            "schema": {
              "default": false,
              "description": "Return a MutationResult with replayed/warnings instead of the bare task. Defaults to false, so existing callers see no change.",
              "title": "Envelope",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/MutationResult-Output"
                    },
                    {
                      "$ref": "#/components/schemas/Task"
                    }
                  ],
                  "title": "Response Handoff Task Api Tasks  Task Id  Handoff Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handoff Task",
        "tags": [
          "status"
        ]
      }
    },
...
  }
}